mv values to shared modules
